linux-generic:pktio mmap: fix nb_rx overflow
bug:
https://bugs.linaro.org/show_bug.cgi?id=2224

+++ platform/linux-generic/pktio/socket_mmap.c
@@ -154,7 +154,7 @@ static inline unsigned pkt_mmap_v2_rx(pktio_entry_t *pktio_entry,
 	int pkt_len;
 	struct ethhdr *eth_hdr;
 	unsigned i = 0;
-	uint8_t nb_rx = 0;
+	unsigned nb_rx = 0;
 	struct ring *ring;
 	int ret;
